Added convertCommandToReadable.
Had to update several data types to derive Show and Read.
This makes converting a String to a TVG command super easy.
The convertCommandToReadable takes a string like

"* 1 -1 h 5"

and converts it to

"Command '*' 1 (-1) (Line Horizontal 5)"

This latter structure is what we can apply read to to
get a command. The function is messy, I will clean it up
later.